+++ stem/__init__.py
@@ -366,6 +366,32 @@ Library for working with the tor process.
   **DISCARD**     throwing out timeout value from when the network was down
   **RESUME**      resumed calculations to determine the proper timeout
   =============== ===========
+
+.. data:: ConnectionType (enum)
+
+  Purpose for a tor connection. Tor may provide types not in this enum.
+
+  The meaning behind these values is a bit unclear, pending :trac:`10086`.
+
+  =============== ===========
+  ConnectionType  Description
+  =============== ===========
+  **OR**          carrying traffic within the tor network
+  **DIR**         fetching or sending tor descriptor data
+  **EXIT**        carrying traffic between the tor network and an external destination
+  =============== ===========
+
+.. data:: TokenBucket (enum)
+
+  Bucket categories of TB_EMPTY events.
+
+  =============== ===========
+  TokenBucket     Description
+  =============== ===========
+  **GLOBAL**      global token bucket
+  **RELAY**       relay token bucket
+  **ORCONN**      bucket used for OR connections
+  =============== ===========
 """
 
 __version__ = '1.1.0-dev'
@@ -697,4 +723,16 @@ TimeoutSetType = stem.util.enum.UppercaseEnum(
   "SUSPENDED",
   "DISCARD",
   "RESUME",
+)
+
+ConnectionType = stem.util.enum.UppercaseEnum(
+  "OR",
+  "DIR",
+  "EXIT",
+)
+
+TokenBucket = stem.util.enum.UppercaseEnum(
+  "GLOBAL",
+  "RELAY",
+  "ORCONN",
 )